fix: Update unit test
* Fix unit test by using full function/method/object path
* Rename unit to test_unit python file
* Move data information into a specific python file
